backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
psycopg2-binary
dj-database-url
python-dotenv
django-cors-headers